refactor(CQ2-1): split notifier.py (1158 lines) into notifier/ package
Split app/utils/notifier.py into a package with focused modules:
- __init__.py: Notifier class (facade combining mixins) + get_notifier/send_notify
- html_builders.py: EmailHtmlMixin (email template, HTML builders, subjects/text)
- webhook.py: WebhookMixin (webhook sending, payload building, header parsing)
- email_sender.py: EmailSenderMixin (SMTP sending)
- selftest.py: TestHelpersMixin (test_notification + channel testers)

Used mixin pattern to preserve all private method access on Notifier instances
(tests directly call notifier._build_email_dynamic_content etc).

Updated 3 test files to patch new module paths:
- app.utils.notifier.httpx -> app.utils.notifier.webhook.httpx
- app.utils.notifier.smtplib -> app.utils.notifier.email_sender.smtplib
- app.utils.notifier.os -> app.utils.notifier.html_builders.os

Renamed test_helpers.py -> selftest.py to avoid pytest collection collision.
All 1894 tests pass, ruff clean.
