+++ app/utils/notifier/html_builders.py
@@ -1,42 +1,13 @@
-"""
-通知模块 - 支持 Webhook 和邮件通知
-"""
+"""通知邮件 HTML 内容构建（mixin）"""
 
 import os
-import smtplib
-import ssl
-import time
-from datetime import datetime
-from email.mime.multipart import MIMEMultipart
-from email.mime.text import MIMEText
-from email.utils import formatdate
-from typing import Any, Callable, Optional
+from typing import Any, Callable
 
-import httpx
-
-from ..core.config import config_manager
-from ..core.logging import logger
+from ...core.logging import logger
 
 
-class Notifier:
-    """通知管理器"""
-
-    def __init__(self, config_manager):
-        self.config_manager = config_manager
-        self._last_notification_time = {}
-        self._notification_cooldown = 60  # 同一类型通知的冷却时间（秒）
-
-    def _should_send_notification(self, notification_type: str) -> bool:
-        """检查是否应该发送通知（防止通知轰炸）"""
-        current_time = time.time()
-        last_time = self._last_notification_time.get(notification_type, 0)
-
-        if current_time - last_time < self._notification_cooldown:
-            logger.debug(f"通知冷却中，跳过 {notification_type} 类型通知")
-            return False
-
-        self._last_notification_time[notification_type] = current_time
-        return True
+class EmailHtmlMixin:
+    """邮件 HTML/文本/标题 构建相关方法（供 Notifier 组合）"""
 
     def _replace_template_variables(self, template: Any, data: dict[str, Any]) -> Any:
         """递归替换模板中的变量"""
@@ -252,255 +223,6 @@ class Notifier:
     <p style="color: #6c757d; font-size: 12px;">此邮件由 Bangumi-Syncer 自动发送</p>
 </body>
 </html>"""
-
-    def send_notification_by_type(
-        self, notification_type: str, data: dict[str, Any]
-    ) -> None:
-        """
-        根据通知类型发送通知
-
-        Args:
-            notification_type: 通知类型（request_received, bangumi_id_found, mark_success, mark_failed, mark_skipped）
-            data: 通知数据（包含timestamp, user_name, title, season, episode, source等）
-        """
-        # 获取所有启用的webhook配置
-        webhook_configs = self._get_webhook_configs()
-
-        for webhook_config in webhook_configs:
-            # 检查是否启用
-            if not webhook_config.get("enabled", False):
-                continue
-
-            # 检查是否支持此通知类型
-            types = webhook_config.get("types", "")
-            if types != "all" and notification_type not in types:
-                continue
-
-            # 检查冷却时间
-            cooldown_key = f"{webhook_config['id']}_{notification_type}"
-            if not self._should_send_notification(cooldown_key):
-                continue
-
-            # 发送webhook通知
-            self._send_webhook_by_config(webhook_config, notification_type, data)
-
-        # 获取所有启用的邮件配置
-        email_configs = self._get_email_configs()
-
-        for email_config in email_configs:
-            # 检查是否启用
-            if not email_config.get("enabled", False):
-                continue
-
-            # 检查是否支持此通知类型
-            types = email_config.get("types", "")
-            if types != "all" and notification_type not in types:
-                continue
-
-            # 检查冷却时间
-            cooldown_key = f"email_{email_config['id']}_{notification_type}"
-            if not self._should_send_notification(cooldown_key):
-                continue
-
-            # 发送邮件通知
-            self._send_email_by_config(email_config, notification_type, data)
-
-    def _get_webhook_configs(self) -> list:
-        """获取所有webhook配置"""
-        config = self.config_manager.get_config_parser()
-        webhook_configs = []
-
-        for section_name in config.sections():
-            if section_name.startswith("webhook-"):
-                section_config = self.config_manager.get_section(section_name)
-                if section_config.get("url"):  # 必须有URL才有效
-                    webhook_configs.append(section_config)
-
-        return webhook_configs
-
-    def _get_email_configs(self) -> list:
-        """获取所有邮件配置"""
-        config = self.config_manager.get_config_parser()
-        email_configs = []
-
-        for section_name in config.sections():
-            if section_name.startswith("email-"):
-                section_config = self.config_manager.get_section(section_name)
-                if section_config.get("smtp_server"):  # 必须有SMTP服务器才有效
-                    email_configs.append(section_config)
-
-        return email_configs
-
-    def _send_webhook_by_config(
-        self,
-        webhook_config: dict[str, Any],
-        notification_type: str,
-        data: dict[str, Any],
-    ) -> bool:
-        """
-        根据配置发送webhook通知
-
-        Args:
-            webhook_config: webhook配置字典
-            notification_type: 通知类型
-            data: 通知数据
-        """
-        try:
-            url = webhook_config["url"]
-            method = webhook_config.get("method", "POST").upper()
-            headers = self._parse_headers(webhook_config.get("headers", ""))
-            template = webhook_config.get("template", "")
-
-            # 构建载荷
-            payload = self._build_payload_by_type(notification_type, data, template)
-
-            # 发送请求
-            logger.info(f"📤 发送 {notification_type} 通知到: {url}")
-
-            if method == "POST":
-                response = httpx.post(url, json=payload, headers=headers, timeout=10)
-            else:  # GET
-                response = httpx.get(
-                    url,
-                    params=payload if isinstance(payload, dict) else None,
-                    headers=headers,
-                    timeout=10,
-                )
-
-            if response.status_code < 300:
-                logger.info(
-                    f"✅ Webhook通知发送成功，响应状态码: {response.status_code}"
-                )
-                return True
-            else:
-                logger.warning(f"⚠️  Webhook返回非成功状态码: {response.status_code}")
-                return False
-
-        except Exception as e:
-            logger.error(f"❌ Webhook通知发送失败: {str(e)}")
-            return False
-
-    def _send_email_by_config(
-        self, email_config: dict[str, Any], notification_type: str, data: dict[str, Any]
-    ) -> bool:
-        """
-        根据配置发送邮件通知
-
-        Args:
-            email_config: 邮件配置字典
-            notification_type: 通知类型
-            data: 通知数据
-        """
-        try:
-            # 获取配置
-            smtp_server = email_config["smtp_server"]
-            smtp_port = int(email_config.get("smtp_port", 587))
-            smtp_username = email_config["smtp_username"]
-            smtp_password = email_config["smtp_password"]
-            smtp_use_tls = email_config.get("smtp_use_tls", True)
-            from_email = email_config.get("email_from", smtp_username)
-            to_email = email_config.get("email_to", "")
-            email_subject = email_config.get("email_subject", "")
-            email_template_file = email_config.get("email_template_file", "")
-
-            # 验证配置
-            if not from_email:
-                from_email = smtp_username
-            if not to_email:
-                logger.warning(
-                    f"邮件配置 ID={email_config.get('id')} 未配置收件人地址，跳过发送"
-                )
-                return False
-
-            # 构建邮件
-            msg = MIMEMultipart("alternative")
-
-            # 使用自定义标题或根据通知类型生成标题
-            if email_subject:
-                subject = self._replace_template_variables(email_subject, data)
-            else:
-                subject = self._build_email_subject_by_type(notification_type, data)
-
-            msg["Subject"] = subject
-            msg["From"] = from_email
-            msg["To"] = to_email
-            msg["Date"] = formatdate(localtime=True)
-
-            # 将 notification_type 添加到 data 字典中，以便模板可以使用
-            data["notification_type"] = notification_type
-
-            # 生成动态内容
-            data["dynamic_content"] = self._build_email_dynamic_content(
-                notification_type, data
-            )
-
-            # 邮件正文
-            text_content = self._build_email_text_by_type(notification_type, data)
-            html_content = self._load_email_template(email_template_file, data)
-
-            # 添加纯文本部分
-            part1 = MIMEText(text_content, "plain", "utf-8")
-            msg.attach(part1)
-
-            # 添加HTML部分
-            part2 = MIMEText(html_content, "html", "utf-8")
-            msg.attach(part2)
-
-            # 发送邮件
-            smtp_port_int = int(smtp_port)
-
-            if smtp_port_int == 465:
-                # 使用 SSL 连接
-                context = ssl.create_default_context()
-                server = smtplib.SMTP_SSL(
-                    smtp_server, smtp_port_int, timeout=30, context=context
-                )
-                try:
-                    server.set_debuglevel(0)
-                    if smtp_username and smtp_password:
-                        server.login(smtp_username, smtp_password)
-                    server.send_message(msg)
-                    server.quit()
-                except Exception as e:
-                    try:
-                        server.quit()
-                    except Exception:
-                        pass
-                    raise e
-            else:
-                # 使用 STARTTLS 连接
-                server = smtplib.SMTP(smtp_server, smtp_port_int, timeout=30)
-                try:
-                    server.set_debuglevel(0)
-                    if smtp_use_tls:
-                        server.starttls()
-                    if smtp_username and smtp_password:
-                        server.login(smtp_username, smtp_password)
-                    server.send_message(msg)
-                    server.quit()
-                except Exception as e:
-                    try:
-                        server.quit()
-                    except Exception:
-                        pass
-                    raise e
-
-            logger.info(
-                f"✅ 邮件通知发送成功: {to_email} (配置ID={email_config.get('id')})"
-            )
-            return True
-
-        except smtplib.SMTPAuthenticationError as e:
-            logger.error(f"❌ 邮件认证失败 (配置ID={email_config.get('id')}): {str(e)}")
-            return False
-        except smtplib.SMTPException as e:
-            logger.error(f"❌ SMTP 错误 (配置ID={email_config.get('id')}): {str(e)}")
-            return False
-        except Exception as e:
-            logger.error(
-                f"❌ 邮件通知发送失败 (配置ID={email_config.get('id')}): {str(e)}"
-            )
-            return False
 
     def _build_email_subject_by_type(
         self, notification_type: str, data: dict[str, Any]
@@ -903,396 +625,3 @@ IP地址: {data.get("ip", "")}
                     <div class="info-value">{data.get("attempt_count", 0)} / {data.get("max_attempts", 0)}</div>
                 </div>
             </div>"""
-
-    def _build_payload_by_type(
-        self, notification_type: str, data: dict[str, Any], template: str
-    ) -> dict[str, Any]:
-        """
-        根据通知类型构建载荷
-
-        Args:
-            notification_type: 通知类型
-            data: 原始数据
-            template: 自定义模板
-        """
-        # 添加通知类型到数据中
-        data["notification_type"] = notification_type
-
-        # 如果有自定义模板，使用模板
-        if template:
-            try:
-                import json
-
-                template_obj = json.loads(template)
-                return self._replace_template_variables(template_obj, data)
-            except Exception as e:
-                logger.warning(f"自定义模板解析失败: {e}，使用默认格式")
-
-        # 根据通知类型使用不同的默认格式
-        default_templates = {
-            "request_received": {
-                "title": "📥 收到同步请求",
-                "type": notification_type,
-                "timestamp": data.get("timestamp", ""),
-                "user": data.get("user_name", ""),
-                "anime": data.get("title", ""),
-                "episode": f"S{data.get('season', 0):02d}E{data.get('episode', 0):02d}",
-                "source": data.get("source", ""),
-            },
-            "bangumi_id_found": {
-                "title": "🔍 匹配到Bangumi番剧",
-                "type": notification_type,
-                "timestamp": data.get("timestamp", ""),
-                "user": data.get("user_name", ""),
-                "anime": data.get("title", ""),
-                "episode": f"S{data.get('season', 0):02d}E{data.get('episode', 0):02d}",
-                "source": data.get("source", ""),
-                "subject_id": data.get("subject_id", ""),
-            },
-            "mark_success": {
-                "title": "✅ 同步成功",
-                "type": notification_type,
-                "timestamp": data.get("timestamp", ""),
-                "user": data.get("user_name", ""),
-                "anime": data.get("title", ""),
-                "episode": f"S{data.get('season', 0):02d}E{data.get('episode', 0):02d}",
-                "source": data.get("source", ""),
-                "subject_id": data.get("subject_id", ""),
-                "episode_id": data.get("episode_id", ""),
-            },
-            "mark_failed": {
-                "title": "❌ 同步失败",
-                "type": notification_type,
-                "timestamp": data.get("timestamp", ""),
-                "user": data.get("user_name", ""),
-                "anime": data.get("title", ""),
-                "episode": f"S{data.get('season', 0):02d}E{data.get('episode', 0):02d}",
-                "source": data.get("source", ""),
-                "error": data.get("error_message", ""),
-                "error_type": data.get("error_type", ""),
-            },
-            "mark_skipped": {
-                "title": "⏭️ 已看过，跳过",
-                "type": notification_type,
-                "timestamp": data.get("timestamp", ""),
-                "user": data.get("user_name", ""),
-                "anime": data.get("title", ""),
-                "episode": f"S{data.get('season', 0):02d}E{data.get('episode', 0):02d}",
-                "source": data.get("source", ""),
-                "subject_id": data.get("subject_id", ""),
-                "episode_id": data.get("episode_id", ""),
-            },
-            "config_error": {
-                "title": "⚠️ 配置错误",
-                "type": notification_type,
-                "timestamp": data.get("timestamp", ""),
-                "error_message": data.get("error_message", ""),
-                "config_type": data.get("config_type", ""),
-                "user_name": data.get("user_name", ""),
-                "mode": data.get("mode", ""),
-            },
-            "anime_not_found": {
-                "title": "🔍 未找到番剧",
-                "type": notification_type,
-                "timestamp": data.get("timestamp", ""),
-                "user": data.get("user_name", ""),
-                "ori_title": data.get("ori_title", ""),
-                "season": data.get("season", 0),
-                "source": data.get("source", ""),
-                "search_method": data.get("search_method", ""),
-            },
-            "episode_not_found": {
-                "title": "📺 未找到剧集",
-                "type": notification_type,
-                "timestamp": data.get("timestamp", ""),
-                "user": data.get("user_name", ""),
-                "season": data.get("season", 0),
-                "episode": data.get("episode", 0),
-                "subject_id": data.get("subject_id", ""),
-                "source": data.get("source", ""),
-            },
-            "api_auth_error": {
-                "title": "🔑 API认证失败",
-                "type": notification_type,
-                "timestamp": data.get("timestamp", ""),
-                "username": data.get("username", ""),
-                "status_code": data.get("status_code", 0),
-                "error_message": data.get("error_message", ""),
-            },
-            "api_error": {
-                "title": "🌐 API错误",
-                "type": notification_type,
-                "timestamp": data.get("timestamp", ""),
-                "status_code": data.get("status_code", 0),
-                "url": data.get("url", ""),
-                "method": data.get("method", ""),
-                "error_message": data.get("error_message", ""),
-                "retry_count": data.get("retry_count", 0),
-            },
-            "api_retry_failed": {
-                "title": "🔄 API重试失败",
-                "type": notification_type,
-                "timestamp": data.get("timestamp", ""),
-                "subject_id": data.get("subject_id", ""),
-                "episode_id": data.get("episode_id", ""),
-                "max_retries": data.get("max_retries", 0),
-                "error_message": data.get("error_message", ""),
-            },
-            "ip_locked": {
-                "title": "🔒 IP被锁定",
-                "type": notification_type,
-                "timestamp": data.get("timestamp", ""),
-                "ip": data.get("ip", ""),
-                "locked_until": data.get("locked_until", ""),
-                "attempt_count": data.get("attempt_count", 0),
-                "max_attempts": data.get("max_attempts", 0),
-            },
-        }
-
-        return default_templates.get(
-            notification_type,
-            {
-                "title": f"📢 {notification_type}",
-                "type": notification_type,
-                "timestamp": data.get("timestamp", ""),
-                "data": data,
-            },
-        )
-
-    def _parse_headers(self, headers_str: str) -> dict[str, str]:
-        """解析请求头字符串"""
-        headers = {"User-Agent": "Bangumi-Syncer-Notifier"}
-
-        # 确保headers_str是字符串类型
-        if not headers_str:
-            return headers
-
-        # 如果headers_str不是字符串，转换为字符串
-        if not isinstance(headers_str, str):
-            headers_str = str(headers_str)
-
-        try:
-            # 尝试解析为JSON
-            import json
-
-            parsed = json.loads(headers_str)
-            if isinstance(parsed, dict):
-                headers.update(parsed)
-        except:
-            # 如果不是JSON，尝试解析为逗号分隔的键值对
-            try:
-                for header in headers_str.split(","):
-                    if ":" in header:
-                        key, value = header.split(":", 1)
-                        headers[key.strip()] = value.strip()
-            except Exception as e:
-                logger.warning(f"解析headers失败: {e}")
-
-        return headers
-
-    def test_notification(
-        self,
-        notification_type: Optional[str] = None,
-        webhook_id: Optional[int] = None,
-        email_id: Optional[int] = None,
-    ) -> dict[str, Any]:
-        """
-        测试通知功能
-
-        Args:
-            notification_type: 通知类型，可选值: 'webhook', 'email', 'all'
-            webhook_id: 指定测试的webhook ID（可选）
-            email_id: 指定测试的email ID（可选）
-
-        Returns:
-            测试结果字典
-        """
-        results: dict[str, Any] = {
-            "webhook": {
-                "enabled": False,
-                "success": False,
-                "message": "",
-                "webhooks": [],
-            },
-            "email": {"enabled": False, "success": False, "message": "", "emails": []},
-        }
-
-        test_data = self._build_test_notification_data()
-
-        if notification_type in (None, "webhook", "all"):
-            self._test_webhook_channels(results, test_data, webhook_id)
-
-        if notification_type in (None, "email", "all"):
-            self._test_email_channels(results, test_data, email_id)
-
-        return results
-
-    @staticmethod
-    def _build_test_notification_data() -> dict[str, Any]:
-        """构建测试通知数据"""
-        return {
-            "timestamp": datetime.now().strftime("%Y-%m-%d %H:%M:%S"),
-            "user_name": "TestUser",
-            "title": "测试番剧",
-            "season": 1,
-            "episode": 1,
-            "source": "test",
-            "error_message": "这是一条测试通知",
-        }
-
-    def _test_webhook_channels(
-        self,
-        results: dict[str, Any],
-        test_data: dict[str, Any],
-        webhook_id: Optional[int],
-    ) -> None:
-        """测试所有 webhook 通道"""
-        webhook_configs = self._get_webhook_configs()
-        if webhook_id:
-            webhook_configs = [w for w in webhook_configs if w.get("id") == webhook_id]
-
-        for webhook_config in webhook_configs:
-            webhook_result: dict[str, Any] = {
-                "id": webhook_config.get("id"),
-                "url": webhook_config.get("url", ""),
-                "enabled": webhook_config.get("enabled", False),
-                "success": False,
-                "message": "",
-            }
-            if webhook_config.get("enabled", False):
-                webhook_result["enabled"] = True
-                try:
-                    success = self._send_webhook_by_config(
-                        webhook_config, "mark_success", test_data
-                    )
-                    if success:
-                        webhook_result["success"] = True
-                        webhook_result["message"] = (
-                            f"Webhook {webhook_config['id']} 测试成功"
-                        )
-                    else:
-                        webhook_result["message"] = (
-                            f"Webhook {webhook_config['id']} 测试失败"
-                        )
-                except Exception as e:
-                    webhook_result["message"] = (
-                        f"Webhook {webhook_config['id']} 测试失败: {str(e)}"
-                    )
-            else:
-                webhook_result["message"] = f"Webhook {webhook_config['id']} 未启用"
-
-            results["webhook"]["webhooks"].append(webhook_result)
-
-        results["webhook"]["enabled"] = len(webhook_configs) > 0
-        results["webhook"]["message"] = f"测试了 {len(webhook_configs)} 个webhook"
-
-    def _test_email_channels(
-        self,
-        results: dict[str, Any],
-        test_data: dict[str, Any],
-        email_id: Optional[int],
-    ) -> None:
-        """测试所有邮件通道"""
-        email_configs = self._get_email_configs()
-        if email_id:
-            email_configs = [e for e in email_configs if e.get("id") == email_id]
-
-        for email_config in email_configs:
-            email_result: dict[str, Any] = {
-                "id": email_config.get("id"),
-                "email_to": email_config.get("email_to", ""),
-                "enabled": email_config.get("enabled", False),
-                "success": False,
-                "message": "",
-            }
-            if email_config.get("enabled", False):
-                email_result["enabled"] = True
-                try:
-                    success = self._send_email_by_config(
-                        email_config, "mark_success", test_data
-                    )
-                    if success:
-                        email_result["success"] = True
-                        email_result["message"] = (
-                            f"邮件配置 {email_config['id']} 测试成功"
-                        )
-                    else:
-                        email_result["message"] = (
-                            f"邮件配置 {email_config['id']} 测试失败"
-                        )
-                except Exception as e:
-                    email_result["message"] = (
-                        f"邮件配置 {email_config['id']} 测试失败: {str(e)}"
-                    )
-            else:
-                email_result["message"] = f"邮件配置 {email_config['id']} 未启用"
-
-            results["email"]["emails"].append(email_result)
-
-        results["email"]["enabled"] = len(email_configs) > 0
-        results["email"]["message"] = f"测试了 {len(email_configs)} 个邮件配置"
-
-
-# 全局通知器实例（延迟初始化）
-_notifier_instance: Optional[Notifier] = None
-
-
-def get_notifier():
-    """获取通知器实例"""
-    global _notifier_instance
-    if _notifier_instance is None:
-        _notifier_instance = Notifier(config_manager)
-    return _notifier_instance
-
-
-def send_notify(
-    notification_type: str, item=None, source: str = None, **kwargs
-) -> bool:
-    """
-    安全发送通知的便捷函数
-
-    Args:
-        notification_type: 通知类型
-        item: CustomItem 对象或 None，自动提取 user_name, title, ori_title, season, episode
-              如果为 None 或字段不存在，使用默认值
-        source: 来源（覆盖 item.source）
-        **kwargs: 额外的通知数据字段，会覆盖从 item 提取的同名字段
-
-    Returns:
-        bool: 是否发送成功
-    """
-    try:
-        notifier = get_notifier()
-
-        # 基础数据，带默认值
-        data = {
-            "timestamp": time.strftime("%Y-%m-%d %H:%M:%S"),
-            "user_name": "unknown",
-            "title": "unknown",
-            "ori_title": "",
-            "season": 0,
-            "episode": 0,
-            "source": "",
-        }
-
-        # 从 item 对象提取字段（安全访问）
-        if item is not None:
-            data["user_name"] = getattr(item, "user_name", "unknown")
-            data["title"] = getattr(item, "title", "unknown")
-            data["ori_title"] = getattr(item, "ori_title", "") or ""
-            data["season"] = getattr(item, "season", 0)
-            data["episode"] = getattr(item, "episode", 0)
-            data["source"] = getattr(item, "source", "")
-
-        # source 参数覆盖
-        if source is not None:
-            data["source"] = source
-
-        # kwargs 覆盖所有同名字段
-        data.update(kwargs)
-
-        notifier.send_notification_by_type(notification_type, data)
-        return True
-    except Exception as e:
-        logger.error(f"发送 {notification_type} 通知失败: {e}")
-        return False
